+++ CHANGELOG.md
@@ -4,6 +4,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## [3.0.2] - 2021-12-24
+
+### Updated
+ 
+- Fix mobile navigation
+
 ## [3.0.1] - 2021-12-22
 
 ### Updated